Add finally block to run()
Wrap most of run() inside a try..finally block, to ensure that the loglist gets
purged regardless of whether that row's import failed or was skipped. Also, change
the test for logger based on an existence check.

One of the logger checks is unnecessary, since we just don't output what's in infolist
if no logger instance exists.

Also add logging if a duplicate was found, and the row was skipped.